requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.24.0

# Try autogen-ext first, fall back to pyautogen if needed
autogen-agentchat>=0.7.4
autogen-core>=0.7.4
autogen-ext>=0.7.4